File and Window Menus
Implimented File and Windows Menus for the Main Control Window (I have yet to figure out a way to create a universal menu bar).

Under File there is "Set Bridge Address…", which affects the program. and "Set User Name…", which currently has no effect.

Under Windows there is an item for each of the four windows with a numberical (ie: ⌘1) shortcut.