Implimented Setting Bridge User Name
Created the procedure setUserName!, which communcates with the Hue Bridge to create a randomly generated user name, as per the recommendation from Philips. This name is stored in a preference file and persists across uses.

Errors (primarily not pressing the link button) are also handled and communicated to the user. The UI for this is still wonky and needs to be fixed. It's in the saveUserName button.

Still tudu: replace the hard coded "brucelighting" user name with the generated user name.