Update UI
Added a menu item for updating the Hue Bridge Firmware. There may be a bug in the update procedure in which it never will execute the updating code. It depends on whether the Bridge automatically updates /swupdate/devicetypes/bridge. I am unable to test this until Philips releases a new version of the bridge software.