Parents & Children
Figured out how to assign a parent object without using a "set-parent" method. (With a begin clause at the end of a class definition). Changed the cue% and light% classes to reflect this. Changed the "Save Cue" button in "Simple Hue Control Main" to assign parent in the creation of a cue and not by using a method.