Change goLights to get state from light% objects
goLights now gets the state to send to the bridge from light% objects. Those states are now compared to the state of the affliated bulb* in the bridge. This avoids the issues that led to only being able to set the lightingState once per use and creates performative json commands.

* Patching is also now possible, though the UI needs to be created.