Implement logic to insertions to a cue-list%
Changed several variables to Racket Style (ie cueList% to cue-list%).

Implimented cue insertion to a cue-list% object by adjusting the the "begin" clause of the cue% class. It checks whether the cue number already exists. If not, it appends the cue to the end of the cue-list% object. If it does exist, it splits the list of children, inserts the new cue% object, and renumbers the cue% objects in the second portion of the cue-list%.

Note, this does not allow for skipping cue numbers while programming. That logic will have to be added later.