Gui side of Inserting cue% objects
Renames several windows, fields, etc to Racket style (ie saveCueDialog to save-cue-dialog).

Adjusts the callback procedure of the save-cue-ok button to impliment the new insertion ability of the cue% class. Also, reorders the cueChoice choice with the new cue-list% children order. This reflects the insertion of the cue% object.

Some of this logic may want to be pulled out into a module.