Spring Cleaning
Deleted the global variables lightsToCue and lightState. Moved global variable bridgeError to start of file. It will eventually be replaced by direct logging. Refactored portions of the main window to be ready to work with a variable number of lights.

Also, redid line breaks to keep the column length to 102 characters (Racket Style).