More Minor Housekeeping
Changed the way the variable setup-needed works to be more intuitive (ie: it prompts the dialog when #t instead of #f). Also, added a check for whether the Bridge Address and User Name are set.

Also, made restoring a cue happen on a separate thread. This needs to be tested.

Finally, changed the names of cue-set-panel and cue-set-button to lights-set-panel and lights-set-button to reflect that they are not related to cues. As for as I know, no one else is working on this program, so it shouldn't be a problem.